refactor(db): firestoreCtrl split **ONLY BACKEND**
