+++ app/src/viewmodels/home/HomeScreenViewModel.tsx
@@ -112,6 +112,10 @@ export function useHomeScreenViewModel(
     }
   }, [user.uid, firestoreCtrl]);
 
+  useEffect(() => {
+
+  })
+
   // Filter challenges to only include those from friends
   const challengesFromFriends = challenges.filter((challenge) =>
     user.friends?.includes(challenge.uid),
